% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateVar.R
\name{aggregateVar}
\alias{aggregateVar}
\title{Extract variable from an object returend by \code{\link{geotopLookUpTable}}}
\usage{
aggregateVar(
  x,
  InputVar = NULL,
  aggregate = c("hourly", "daily", "monthly", "yearly"),
  aggregate_fun = c("mean", "sum"),
  ...
)
}
\arguments{
\item{x}{objectect returned by \code{\link{geotopLookUpTable}} with \code{merge==TRUE}}

\item{InputVar}{discover variable}

\item{aggregate}{aggregate option. Default it is \code{c("hourly","daily","monthly","yearly")} and it is considered the first element.}

\item{aggregate_fun}{aggregate function for \code{InputVar} and \code{Add_InputVar} respectively. Deafault is \code{c("mean","sum")}.}

\item{...}{further arguments}
}
\description{
Extract variable from an object returend by \code{\link{geotopLookUpTable}}
}
\note{
It is assumed that \code{x} already contains hourly aggregated time series, so in case \code{aggregare=="hourly"} no aggragetion is calculated.
}
\examples{
wpath <- system.file('geotop-simulation/B2_BeG_017',package="geotopOptim2")
wpath <- system.file('geotop-simulation/B2site',package="geotopOptim2")
ex <- geotopLookUpTable(wpath = wpath)
 
eout <- aggregateVar(ex)
dout <- aggregateVar(ex,aggregate="daily")
mout <-  aggregateVar(ex,aggregate="monthly")



}
\seealso{
\code{\link{geotopLookUpTable}}
}
