% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxfunDataFrame.R
\name{approxfunDataFrame}
\alias{approxfunDataFrame}
\title{Interpolation on the soil layers}
\usage{
approxfunDataFrame(df, z = NULL, zout, formatter = "z\%04d", factor = 10,
  ...)
}
\arguments{
\item{df}{data frame or 'zoo' object with (modeled) time series.}

\item{z}{vector of soil layer depths. Default is \code{NULL} and is taken from column names of \code{df}.}

\item{zout}{vector of soil layer depth at which \code{df} values are interpoleted with \code{\link{approx}}}

\item{formatter}{character string with decimal formatter. It is used if \code{z} or {zout} are string vectors. Default is \code{"z\%04d"}.}

\item{factor}{unt factor used for conversion of  \code{z} or {zout} from String vector to numeric vector and viceversa.}

\item{...}{further argument for \code{\link{approx}}}
}
\value{
The values of the \code{df} quantity interpolated in the \code{zout} soil depths.
}
\description{
This function interpolates the values of a \code{zt}-variable quantity in layers of a carteain depth.
}
\examples{
## NOTHING
}
\seealso{
\code{\link{approx}}
}

