% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.geotop.map.value.R
\name{extract.geotop.value.fromMap}
\alias{extract.geotop.value.fromMap}
\title{Getting values from a GEOtop Simulation Map}
\usage{
extract.geotop.value.fromMap(key, xykeys = c("PointFile",
  "HeaderCoordinatePointX", "HeaderCoordinatePointY"), wpath,
  inpts.file = "geotop.inpts", ...)
}
\arguments{
\item{key}{keywords related to the maps}

\item{xykeys}{keywords keywords relate to the selected chck points where to extract the values. Default is \code{c("Listpoint","HeaderCoordinatePointX","HeaderCoordinatePointY")}, where the first keyword refers to the file with XY Coordinates of the check points, the second and the third ones refer to  the headers of X and Y Coordinates respectively.}

\item{wpath, inpts.file, ...}{arguments that are passed to \code{\link{get.geotop.inpts.keyword.value}}}
}
\description{
This function extracts the values of one or more GEOtop maps in certain check point.
}
\examples{

wpath <- "/Users/ecor/Dropbox/R-packages/geotopOptim-stuff-SHALINI/exercise/simulation_template"

points <- extract.geotop.value.fromMap(key="SoilMapFile",wpath=wpath)
points0 <- extract.geotop.value.fromMap(key=c("SoilMapFile","DemFile"),wpath=wpath)

nopoints <- extract.geotop.value.fromMap(key="NoSoilMapFile",wpath=wpath)
nopoints0 <- extract.geotop.value.fromMap(key=c("NoSoilMapFile","DemFile"),wpath=wpath)
nopoints1 <- extract.geotop.value.fromMap(key=c("NoSoilMapFile","NoDemFile"),wpath=wpath)

}
\seealso{
\code{\link{get.geotop.inpts.keyword.value}}
}

