% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotop.execution.R
\name{geotopExec}
\alias{geotopExec}
\title{Execution of GEOtop Hydrological Model}
\usage{
geotopExec(
  param = NULL,
  bin = "/home/ecor/local/geotop/GEOtop/bin/geotop-2.0.0",
  simpath,
  inpts.file = "geotop.inpts",
  runpath = "/home/ecor/temp/geotopOptim_tests",
  temporary.runpath = FALSE,
  clean = TRUE,
  recovery = !clean,
  getKeywords = NULL,
  data.frame = TRUE,
  date_field = "Date12.DDMMYYYYhhmm.",
  intern = FALSE,
  param.soil = TRUE,
  formatter = "\%04d",
  paramPrefix = "Header",
  names_par = NULL,
  SoilType = 1,
  level = 1,
  time.aggregate = NULL,
  ...
)
}
\arguments{
\item{param}{vector of parameters to set for GEOtop simulation. Actually implemented only for soil type parameters. Deafault is \code{NULL}.}

\item{bin}{binary executable file of GEOtop with full path}

\item{simpath}{directory containg the \code{inpts.file}.}

\item{inpts.file}{GEOtop configuretion \code{*.inpts} file. Default is \code{geotop.inpts}.}

\item{runpath}{directory where to run GEOtop}

\item{temporary.runpath}{logical value. If it \code{TRUE} GEOtop is running in a sub-directory of \code{runpath} made with randomly generated name.  Default is \code{FALSE}.}

\item{clean}{logical value. If it is \code{TRUE} previous simulations and other stuff in \code{runpath} were removed. See \code{\link{file.remove}} for functionality.}

\item{recovery}{logical value. If it is \code{TRUE}, GEOtop simulation are recovered or overwriiten according to the GEOtop settings in \code{inpts.file}.}

\item{getKeywords}{character vector containing the keywords of \code{inpts.file} which can be imported after the GEOtop run. Default is \code{NULL}, nothing is imported.}

\item{data.frame}{logical vaue, see  \code{\link{get.geotop.inpts.keyword.value}}.  Default is \code{TRUE}, is enabled if \code{getKeywords} is not \code{NULL}.}

\item{date_field}{character vaue, see  \code{\link{get.geotop.inpts.keyword.value}}. Default is \code{"Date12.DDMMYYYYhhmm."}, is enabled if \code{getKeywords} is not \code{NULL}.}

\item{intern}{logical value, see \code{\link{system}}. OBSOLETE}

\item{param.soil}{logical value If it \code{TRUE} (default) the vaues in \code{param} concerns soil parameters.}

\item{formatter}{character decimal formatter.  See \code{\link{get.geotop.inpts.keyword.value}}. Default is \code{"\%04d"}.}

\item{paramPrefix}{character string. Default is \code{"Header"}. If \code{param.soil==TRUE}, the soil parameters in \code{param} are named with the corrosponding parameter keywords in \code{inpts.file} to which the  string \code{paramPrefix} is attached as a prefix. If \code{paramPrefix} is \code{NA} or \code{NULL} , the names of \code{param} elements must be matched with the corresponimg paramater name in the soil parameter data frame.}

\item{names_par}{OPTIONAL vector of names for \code{param}}

\item{SoilType}{soil type used for GEOtop Calibration. This must be an ID of the soil types indicated in the \code{inpts.file} of The GEOtop Simulation. Default is 1.  If it is \code{NA} or \code{NULL}, the soil type is extracted from the soil map (\code{SoilMapFile}).}

\item{level}{ID check point of a distributed or quasi-distributed GEOtop simulatiion where to perform the Point Calibration. Default is 1. See \code{\link{get.geotop.inpts.keyword.value}}}

\item{time.aggregate}{list of the arguments for \code{\link{aggregate.zoo}}. If it is \code{NULL} (default) no time aggragation is applied to the function outputs. Otherwise, the outputs are aggregated within time intervals. In case of monthly averaged values, it must  be set as \code{list(FUN=mean,by="\%Y-\%m",na.rm=TRUE)} where the \code{by} item is the string format according to POSIX standard for times and dates represented the aggregation interval ( sse \code{\link{strptime}}.}

\item{...}{further arguments for \code{\link{get.geotop.inpts.keyword.value}}}
}
\description{
Execution of GEOtop Hydrological Model
}
\details{
In this function, implementation, The parameters entered through \code{param} replace  only the ones of  the first (\code{"0001"}) soil type  declared in the GEOtop simulation directory, if \code{SoilType} argument is omitted. The elements of the vector \code{param} must be named with the respective keywords preceded by the prefix \code{SOIL__} if the parameter is a soil property through \code{SoilParFIle} file (see package documentation); \code{SCALAR__} if the parameter is scalar (one numeric value) and assigned directly in the \code{inpts.file} file (e.g. \code{"geotop.inpts"}); \code{VECTOR__[X]__} if the parameter is assigned in the \code{inpts.file} file and is the \code{X}-th element of a vector (e.g. assignments of land use proprties).
}
\examples{

simpath <- system.file('geotop-simulation/B2site',package="geotopOptim2")  #######system.file("Muntatschini_pnt_1_225_B2_004",package="geotopOptim")
bin <-   "/home/ecor/local/geotop/GEOtop/bin/geotop-2.0.0"
runpath <- "/home/ecor/temp/geotopOptim_tests"

vars <- c("SoilAveragedTempProfileFile",	"SoilLiqWaterPressProfileFile",
"SoilLiqContentProfileFile","SoilIceContentProfileFile",
"AvailableSoilWaterContent","PointOutputFile")


out <- geotopExec(bin=bin,simpath=simpath,runpath=runpath,
      clean=TRUE,getKeywords=vars,data.frame=TRUE,level=1)

param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05) 
out1 <- geotopExec(param=param,bin=bin,simpath=simpath,
			runpath=runpath,clean=TRUE,getKeywords=vars,
			data.frame=TRUE,level=1,intern=TRUE)


 ## Monthly-Aggregated Results: 

 out2 <- geotopExec(param=param,bin=bin,simpath=simpath,
			runpath=runpath,clean=TRUE,getKeywords=vars,
			data.frame=TRUE,level=1,intern=TRUE,
			time.aggregate=list(FUN=mean,by="\%Y-\%m",na.rm=TRUE))



 ## SCALAR__ and VECTOR__ input params

param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05,SCALAR__LSAI=4.2)

out3 <- geotopExec(param=param,bin=bin,simpath=simpath,
			runpath=runpath,clean=TRUE,getKeywords=vars,
			data.frame=TRUE,level=1,intern=TRUE)


param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05,VECTOR__1__LSAI=4.2)

out4 <- geotopExec(param=param,bin=bin,simpath=simpath,
			runpath=runpath,clean=TRUE,getKeywords=vars,
			data.frame=TRUE,level=1,intern=TRUE)





}
