% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotop.gof.2.R
\name{geotopGOF}
\alias{geotopGOF}
\title{GEOtop Goodnes of fit}
\usage{
geotopGOF(x = NULL, run.geotop = TRUE, target = NULL, uscale = NULL,
  when = NULL, gof.mes = "RMSE", np = 2, nosuccess.return = Inf, ...)
}
\arguments{
\item{x}{vector with parameters to calibrate. See \code{param} in \code{\link{geotopExec}}.}

\item{run.geotop}{logical value. Default is \code{TRUE}. It it is \code{TRUE} GEOtop is runned though \code{\link{geotopExec}}.}

\item{target}{variables used for comparasion. The name of the variable must be in accordance with observation names returned by \code{\link{geotopLookUpTable}}.}

\item{uscale}{unit scale value for each target variable. (...)}

\item{when}{vector of time instants (class \code{\link{POSIXct}}. It is passed as \code{when} to \code{\link{geotopLookUpTable}}. If it is \code{NULL} all observation/simulated time duration is considered.}

\item{gof.mes}{measure (index) of godness of fit. The measure must be one of the row names of \code{\link{gof}}}

\item{np}{exponent for te p-norm used. It must be an integer greater or equal to 1. It can be 1 (Manhattan), 2 (Euclidean) or \code{Inf} (Component of the Maximum absoulte value).  Default is 2.}

\item{nosuccess.return}{value thet function returns in case of non-0 (e.g. 1) exit of GEOtop simulation. Default is \code{Inf}.}

\item{...}{further arguments (see \code{\link{geotopExec}} and \code{\link{gof}}.}
}
\description{
GEOtop Goodnes of fit
}
\examples{

wpath <- system.file('geotop-simulation/B2site',package="geotopOptim2")
bin <-   "/home/ecor/local/geotop/GEOtop/bin/geotop-2.0.0"
runpath <- "/home/ecor/temp/geotopOptim_tests"

var <- 'soil_moisture_content_50'

param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05) 
ssout <- geotopGOF(x=param,run.geotop=TRUE,bin=bin,
			simpath=wpath,runpath=runpath,clean=TRUE,data.frame=TRUE,
			level=1,intern=TRUE,temporary.runpath=TRUE)


#
}
\seealso{
\code{\link{geotopExec}},\code{\link{gof}}
}

