% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotop.pso.2.R
\name{geotopPSO}
\alias{geotopPSO}
\alias{geotoplhoat}
\title{GEOtop calibration through Particle Swam Optimization}
\usage{
geotopPSO(par = NULL, fn = geotopGOF, gof.mes = "KGE", final.run = TRUE,
  upper, lower, ..., hydroPSOfun = c("hydroPSO", "lhoat"))

geotoplhoat(...)
}
\arguments{
\item{par}{model parameters. ((TO DO) See \code{\link{hydroPSO}},\code{\link{geotopGOF}} and \code{\link{geotopExec}}}

\item{fn}{function to optimize (minimize or maximize). Default is \code{\link{geotopGOF}}. See \code{\link{hydroPSO}}.}

\item{gof.mes}{string(s) containing adopted numerical goodness-of-fit measure. If it is \code{NULL} (Default), all mesasures returned by \code{\link{gof}} are calculated.}

\item{final.run}{logical value. It is \code{TRUE} (default), simulated time series with optimal set of parameteers are added in the list object returned by the function.}

\item{upper, lower}{see  \code{upper} and \code{lowe} in \code{\link{hydroPSO}}}

\item{...}{further arguments for \code{\link{hydroPSO}}.}

\item{hydroPSOfun}{used function name of \code{hydroPSO} package: \code{\link{hydroPSO}} or \code{\link{lhoat}}.}
}
\description{
This function performs a calibration or a senisitivity analyisis of 
the GEOtop Distributed Hydrological Model through a Particle Swam Optimization or LH-OAT method respectively.
This function is a wrapper of \code{\link{hydroPSO}} or \code{\link{lhoat}}. 
The use of either  \code{\link{hydroPSO}} or \code{\link{lhoat}} is declared by the argument \code{hydroPSOfun}.
}
\details{
The function \code{fn}, in case it is different from the default value \code{\link{geotopGOF}} , must always have the arguments \code{gof.mes} and \code{gof.expected.value.for.optim}.
}
\examples{



wpath <- system.file('geotop-simulation/B2site',package="geotopOptim2") ####'/home/ecor/activity/2016/eurac2016/idra/B2_BeG_017_DVM_001_test_1'
bin <-   "/home/ecor/local/geotop/GEOtop/bin/geotop-2.0.0"
runpath <- "/home/ecor/temp/geotopOptim_tests"

var <- 'soil_moisture_content_50'

param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05) 
ssout <- geotopGOF(x=param,run.geotop=TRUE,bin=bin,
			simpath=wpath,runpath=runpath,clean=TRUE,data.frame=TRUE,
			level=1,intern=TRUE,target=var,temporary.runpath=TRUE)





x <- param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05,
  LateralHydrConductivity=0.021,NormalHydrConductivity=0.021) 
upper <- x*3

upper["LateralHydrConductivity"] <- 0.1
upper["NormalHydrConductivity"] <- 0.1

lower <- x/3
lower["N"] <- 1.1
lower["LateralHydrConductivity"] <- 0
lower["NormalHydrConductivity"] <- 0


control <- list(maxit=4,npart=2) ## Maximim 4 iterations!! 

\dontrun{
pso <- geotopPSO(par=x,run.geotop=TRUE,bin=bin,
			simpath=wpath,runpath=runpath,clean=TRUE,data.frame=TRUE,
			level=1,intern=TRUE,target=var,gof.mes="KGE",lower=lower,upper=upper,control=control)
}
#####lhoat <- geotoplhoat(par=x,obs=obs_SWC,... ## TO DO

}
\seealso{
\code{\link{hydroPSO}},\code{\link{gof}}\code{\link{lhoat}}
}

