% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.wrapper.R
\name{gofg}
\alias{gofg}
\title{Wrapper function for hydroGOF::gof function}
\usage{
gofg(sim, obs, ...)
}
\arguments{
\item{sim, obs, ...}{argument for \code{\link{gof}}}
}
\description{
Wrapper function for hydroGOF::gof function
}
\details{
This functio is a wrapper of \code{\link{gof}}. In case \code{obs} or \code{sim} are all \code{NA}, it returns a matrix of \code{NA}.
}
\seealso{
\code{\link{gof}}
}
