% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratefunDataFrame.R
\name{integratefunDataFrame}
\alias{integratefunDataFrame}
\title{Integration on the soil layers}
\usage{
integratefunDataFrame(
  df,
  z = NULL,
  zrange = c(0, 500),
  formatter = "z\%04d",
  factor = 10,
  rescaleWithDz = FALSE,
  ...
)
}
\arguments{
\item{df}{data frame or 'zoo' object with (modeled) time series.}

\item{z}{vector of soil layer depths. Default is \code{NULL} and is taken from column names of \code{df}.}

\item{zrange}{vector of soil layer thickness within which \code{df} values are integrated (or averaged)  with \code{\link{approx}} and then integrated with  \code{\link{integrate}}.}

\item{formatter}{character string with decimal formatter. It is used if \code{z} or {zout} are string vectors. Default is \code{"z\%04d"}.}

\item{factor}{unit factor used for conversion of  \code{z} or {zout} from String vector to numeric vector and viceversa.}

\item{rescaleWithDz}{logical value. If is \code{TRUE} the integrated value is rescaled with soil thickness (mean). Default is \code{FALSE} and the value is not rescaled.}

\item{...}{further argument for \code{\link{approx}}}
}
\value{
The values of the \code{df} quantity interpolated in the \code{zout} soil depths.
}
\description{
This function integrates the values of a \code{zt}-variable quantity  within a certain thickness.
}
\examples{

## NOTHING 
}
\seealso{
\code{\link{approx}},\code{\link{integrate}}
}
