% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateVar.R
\name{aggregateVar}
\alias{aggregateVar}
\title{Extract variable from an object returend by \code{\link{geotopLookUpTable}}}
\usage{
aggregateVar(x, InputVar = NULL, aggregate = c("hourly", "daily", "monthly",
  "yearly"), aggregate_fun = c("mean", "sum"), ...)
}
\arguments{
\item{x}{objectect returned by \code{\link{geotopLookUpTable}} with \code{merge==TRUE}}

\item{InputVar}{discover variable}

\item{aggregate}{aggregate option. Default it is \code{c("hourly","daily","monthly","yearly")} and it is considered the first element.}

\item{aggregate_fun}{aggregate function for \code{InputVar} and \code{Add_InputVar} respectively. Deafault is \code{c("mean","sum")}.}

\item{...}{further arguments}

\item{Add_InputVar}{additional variable (simulated)}
}
\description{
Extract variable from an object returend by \code{\link{geotopLookUpTable}}
}
\note{
It is assumed that \code{x} already contains hourly aggregated time series, so in case \code{aggregare=="hourly"} no aggragetion is calculated.
}
\examples{
wpath <- '/home/ecor/activity/2016/eurac2016/idra/B2_BeG_017_DVM_001_test_1' 


ex <- geotopLookUpTable(wpath = wpath)
 
eout <- aggregateVar(ex)
dout <- aggregateVar(ex,aggregate="daily")
mout <-  aggregateVar(ex,aggregate="monthly")



}
\seealso{
\code{\link{geotopLookUpTable}}
}

